courses/atf_cloud_interactive/setup/gcs-mcp-server/requirements.txt
fastmcp>=2.0
google-auth>=2.0
google-cloud-storage>=2.18
uvicorn>=0.32
